# Conflicts:
#	test/__snapshots__/UniswapV3Factory.spec.ts.snap
#	test/__snapshots__/UniswapV3Pair.spec.ts.snap

+++ contracts/test/MockTimeUniswapV3Pair.sol
@@ -6,7 +6,7 @@ import '../UniswapV3Pair.sol';
 
 // used for testing time dependent behavior
 contract MockTimeUniswapV3Pair is UniswapV3Pair {
-    uint32 public time;
+    uint64 public time;
 
     constructor(
         address factory,
@@ -14,11 +14,11 @@ contract MockTimeUniswapV3Pair is UniswapV3Pair {
         address tokenB
     ) public UniswapV3Pair(factory, tokenA, tokenB) {}
 
-    function setTime(uint32 _time) external {
+    function setTime(uint64 _time) external {
         time = _time;
     }
 
-    function _blockTimestamp() internal view override returns (uint32) {
+    function _blockTimestamp() internal view override returns (uint64) {
         return time;
     }
 }
